requirements.txt
absl-py==2.1.0
altair==5.3.0
appdirs==1.4.4
asttokens==2.4.1
astunparse==1.6.3
attrs==23.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
cmake==3.29.2
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.1.1
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
einops==0.8.0
executing==2.0.1
filelock==3.13.1
flatbuffers==24.3.25
fonttools==4.44.0
frozendict==2.4.3
fsspec==2023.10.0
fvcore==0.1.5.post20221221
gast==0.4.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.29.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.63.0
h5py==3.11.0
html5lib==1.1
huggingface-hub==0.23.3
humanfriendly==10.0
idna==3.7
importlib-resources==6.1.1
importlib_metadata==7.1.0
iopath==0.1.10
ipykernel==6.29.4
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
keras==2.13.1
kiwisolver==1.4.5
libclang==18.1.1
lit==18.1.3
lxml==5.2.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
multitasking==0.0.11
natsort==8.4.0
nest-asyncio==1.6.0
networkx==3.1
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu11==2.14.3
nvidia-nvjitlink-cu12==12.3.52
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.1.105
onnx==1.14.0
onnxruntime==1.15.1
opencv-python==4.10.0.84
opt-einsum==3.3.0
packaging==23.2
pandas==2.0.3
parso==0.8.4
peewee==3.17.3
pickleshare==0.7.5
Pillow==10.1.0
pkgutil_resolve_name==1.3.10
platformdirs==4.2.2
portalocker==2.8.2
prompt_toolkit==3.0.46
protobuf==5.26.1
psutil==5.9.6
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==16.1.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycocotools==2.0.7
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==26.0.3
referencing==0.35.1
regex==2024.5.15
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
safetensors==0.4.3
scikit-learn==1.3.2
scikit-surprise==1.1.4
scipy==1.10.1
seaborn==0.13.0
six==1.16.0
smmap==5.0.1
soupsieve==2.5
stack-data==0.6.3
streamlit==1.35.0
sympy==1.12
tabulate==0.9.0
tenacity==8.3.0
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.1
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.4.0
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
tokenizers==0.19.1
toml==0.10.2
toolz==0.12.1
torch==2.0.1
torchvision==0.15.2
tornado==6.4.1
tqdm==4.66.1
traitlets==5.14.3
transformers==4.41.2
triton==2.0.0
typing_extensions==4.12.1
tzdata==2023.3
ultralytics==8.0.209
urllib3==2.2.1
watchdog==4.0.1
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.0.2
wrapt==1.16.0
yacs==0.1.8
yfinance==0.2.38
zipp==3.17.0
